Updated repo_archiver.py and README.md
During usage, discovered that occasionally issue closing would create a "422 unprocessable entity" error.  Was able to reproduce in the UI, but the closure would stick.

Added a --force flag to allow you to rerun archiver on a partially archived thing
Added error reporting on the 422 error.  Sees it, but keeps going.  Have been unable to reproduce since the addition, so it's unclear if the issue will close like in the UI