Update to repo_archiver and unarchiver to handle custom text labels
In order to better communicate when we do bulk closures, the desire to have custom text added to the description/label of issues/repos is handy - specifically to have a shortened URL embedded that links to "So, you've been archived!" documentation with how to request reversal.  The archiver gets this custom flag, and the unarchiver is agnostic, it just finds and handles it.