Initial add of repo_issue_create.py and example for the associated conf file
Script to open issues based on the issueconf.yaml.example file
title: title of the issue
body: Body of the issue
bodyUID: some unique string for identification of the issue
comment: comment for any already existing issue.

Has options to reopen issues if they're found closed, and add the comment to that or other issues

adding UID to the issues to allow for better identification of issues.